update: 사용자 검색 시 불필요한 중복 형변환 제거

+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/member/repository/impl/MemberExposedRepositoryImpl.kt
@@ -50,7 +50,7 @@ class MemberExposedRepositoryImpl : MemberExposedRepository {
                 .selectAll()
                 .apply { whereClause?.let { where { it } } }
                 .limit(pageable.pageSize)
-                .offset(pageable.offset.toLong())
+                .offset(pageable.offset)
                 .map { row ->
                     Member(
                         id = row[MemberExposedEntity.id],
